refactor: Move `useReferencedContactName` to hooks folder
